requirements.txt
-r lint-requirements.txt
pre-commit>=2.3.0
desed>=1.3.6
pytorch-lightning==1.5.1
asteroid==0.4.1
dcase_util==0.2.16
psds_eval>=0.4.0
sed_eval==0.2.1